Fixed 32bit windows build.

+++ OrbitCore/Utils.h
@@ -524,7 +524,7 @@ std::string FormatTime(const time_t& rawtime);
 
 //-----------------------------------------------------------------------------
 bool ReadProcessMemory(uint32_t pid, uint64_t address, byte* buffer,
-                       uint64_t size, size_t* num_bytes_read);
+                       uint64_t size, uint64_t* num_bytes_read);
 
 #define CONCAT_(x, y) x##y
 #define CONCAT(x, y) CONCAT_(x, y)
